clean splash router
