+++ lib/widgets/pages/splash/splash_loading_page.dart
@@ -1,0 +1,11 @@
+import 'package:coffeecard/base/style/colors.dart';
+import 'package:flutter/material.dart';
+
+class SplashLoadingPage extends StatelessWidget {
+  const SplashLoadingPage();
+
+  @override
+  Widget build(BuildContext context) {
+    return Container(color: AppColor.primary);
+  }
+}
